# docs/OpenApi.yaml
openapi: 3.0.0
info:
  title: BankApp API
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Локальный сервер разработки
...
components:
  securitySchemes:
    cookieAuth:
      $ref: "./components.yaml#/components/securitySchemes/cookieAuth"
  schemas:
    Error:
      $ref: "./components.yaml#/components/schemas/Error"
    SuccessResponse:
      $ref: "./components.yaml#/components/schemas/SuccessResponse"
...
    UserUpdate:
      $ref: "./components.yaml#/components/schemas/UserUpdate"
    UserGet:
      $ref: "./components.yaml#/components/schemas/UserGet"
    AdminRegister:
      $ref: "./components.yaml#/components/schemas/AdminRegister"
    ChangeRoleRequest:
      $ref: "./components.yaml#/components/schemas/ChangeRoleRequest"
...
paths:
...
  /api/v1/basic/user:
    get:
      summary: Получение данных текущего пользователя
      tags:
        - Users
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Успешный ответ с данными пользователя
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/UserGet"
        "400":
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      summary: Обновление данных текущего пользователя
      tags:
        - Users
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Данные пользователя успешно обновлены
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated_user:
                    $ref: "#/components/schemas/UserGet"
        "400":
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Удаление текущего пользователя
      tags:
        - Users
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Пользователь успешно удален
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/admin/users:
    post:
      summary: Создание администратора или верификатора
      tags:
        - Admin
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminRegister"
      responses:
        "201":
          description: Пользователь успешно создан
          content:
            application/json:
              schema:
                type: object
                properties:
                  admin_or_verificator_id:
                    type: string
                    format: uuid
                    example: "b6609ddd-95f4-42f0-993e-7f07f3fa1d5b"
        "400":
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Доступ запрещен (требуется роль Admin)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      summary: Получение списка пользователей с фильтром по роли
      tags:
        - Admin
      security:
        - cookieAuth: []
      parameters:
        - name: role
          in: query
          required: true
          description: Роль пользователей для фильтрации
          schema:
            type: string
            enum: [Basic, Verificator, Admin]
          example: "Verificator"
      responses:
        "200":
          description: Успешный ответ со списком пользователей
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/UserGet"
        "403":
          description: Доступ запрещен (требуется роль Admin)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/admin/users/{user_id}:
    patch:
      summary: Изменение роли пользователя
      tags:
        - Admin
      security:
        - cookieAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          description: Идентификатор пользователя
          schema:
            type: string
            format: uuid
          example: "b6609ddd-95f4-42f0-993e-7f07f3fa1d5b"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeRoleRequest"
      responses:
        "200":
          description: Роль пользователя успешно изменена
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/UserGet"
        "400":
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Доступ запрещен (требуется роль Admin)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...